Rename example_app to gql_platform (more appropriate).
At this stage of usage.
