+++ docker-backend/graphene-django-backend/config/settings/base.py
@@ -56,7 +56,7 @@ THIRD_PARTY_APPS = (
 )
 
 LOCAL_APPS = (
-    'project.example_app',
+    'project.gql_platform',
 )
 
 INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + LOCAL_APPS
