+++ docker-backend/graphene-django-backend/project/gql_platform/admin.py
