+++ docker-backend/graphene-django-backend/project/gql_platform/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class ExampleAppConfig(AppConfig):
-    name = 'example_app'
+    name = 'gql_platform'
