+++ docker-backend/graphene-django-backend/project/gql_platform/models.py
