+++ docker-backend/graphene-django-backend/project/gql_platform/tests/.coverage
@@ -1,0 +1,1 @@
+!coverage.py: This is a private format, don't read it directly!{"lines":{}}
