+++ docker-backend/graphene-django-backend/project/gql_platform/views.py
